feat(#381): MCP Streamable HTTP transport + protocol version bump
Protocol:
  - Bump protocolVersion from 2024-11-05 to 2025-06-18 (current spec)

Architecture:
  - Extract MCP handler into uteke_mcp library crate (lib.rs)
  - main.rs simplified to just stdio transport loop delegating to lib
  - uteke-server gains POST /mcp endpoint that calls the shared handler

HTTP endpoint:
  - POST /mcp accepts JSON-RPC body (initialize, tools/list, tools/call)
  - Returns application/json with MCP-Protocol-Version header
  - Uses existing Mutex<Uteke> from uteke-server (thread-safe)
  - Enables remote MCP clients without subprocess spawning

Closes #381
