+++ crates/uteke-server/Cargo.toml
@@ -15,6 +15,7 @@ path = "src/main.rs"
 
 [dependencies]
 uteke-core = { path = "../uteke-core", version = "0.2.0" }
+uteke-mcp = { path = "../uteke-mcp", version = "0.2.0" }
 serde = { version = "1", features = ["derive"] }
 serde_json = "1"
 toml = "1"
