+++ crates/uteke-server/src/main.rs
@@ -862,6 +862,24 @@ fn route(uteke: &Mutex<Uteke>, ctx: &ReqCtx, req: &mut Request) -> Response<Curs
             Err(e) => ctx.error_response_for(req, 400, e),
         },
 
+        // ── MCP JSON-RPC endpoint (#381) ───────────────────────────────
+        (Method::Post, "/mcp") => {
+            let mut body = String::new();
+            if let Err(e) = req.as_reader().read_to_string(&mut body) {
+                return ctx.error_response_for(req, 400, format!("Failed to read body: {e}"));
+            }
+            let response = uteke_mcp::handle_jsonrpc(&uteke, &body);
+            tiny_http::Response::from_string(response)
+                .with_header(
+                    tiny_http::Header::from_bytes(&b"Content-Type"[..], &b"application/json"[..])
+                        .unwrap(),
+                )
+                .with_header(
+                    tiny_http::Header::from_bytes(&b"MCP-Protocol-Version"[..], &b"2025-06-18"[..])
+                        .unwrap(),
+                )
+        }
+
         // ── 404 ─────────────────────────────────────────────────────────
         _ => ctx.error_response_for(req, 404, "Not found"),
     }
